+++ lib/ui/views/home_view.dart
@@ -3,6 +3,7 @@ import 'package:privates_app/core/decorations/color_palette.dart';
 import 'package:privates_app/core/decorations/device_scaler.dart';
 import 'package:privates_app/generated/l10n.dart';
 import 'package:privates_app/ui/view-models/home_view_model.dart';
+import 'package:privates_app/ui/widgets/add_user_widget.dart';
 import 'package:privates_app/ui/widgets/post_widget.dart';
 import 'package:stacked/stacked.dart';
 
@@ -76,172 +77,12 @@ class HomeView extends StatelessWidget {
                 child: ListView(
                   shrinkWrap: true,
                   scrollDirection: Axis.horizontal,
-                  children: [
-                    Container(
-                      width: 150,
-                      margin: const EdgeInsets.symmetric(horizontal: 10),
-                      padding: EdgeInsets.symmetric(vertical: DeviceScaler().scale(10)),
-                      decoration: BoxDecoration(
-                          color: Palette.white,
-                          borderRadius: BorderRadius.circular(10),
-                          boxShadow: const [
-                            BoxShadow(
-                                blurRadius: 50,
-                                offset: Offset(0,0),
-                                color: Palette.lightShadowGrey
-                            )
-                          ]
-                      ),
-                      child: Column(
-                        mainAxisAlignment: MainAxisAlignment.end,
-                        children: [
-                          Image.asset("assets/png/big-photo.png"),
-                          Text(
-                            "Billy Morgan",
-                            style: TextStyle(
-                              color: Palette.black,
-                              fontSize: DeviceScaler().scaleFont(14),
-                              fontWeight: FontWeight.w500,
-                            ),
-                          ),
-                          SizedBox(height: DeviceScaler().scale(3),),
-                          Text(
-                            "Actress , Life coach, \nphilanthropist",
-                            textAlign: TextAlign.center,
-                            style: TextStyle(
-                              color: Palette.grey,
-                              fontSize: DeviceScaler().scaleFont(10),
-                              fontWeight: FontWeight.w500,
-                            ),
-                          ),
-                          SizedBox(height: DeviceScaler().scale(5),),
-                          Container(
-                            padding: EdgeInsets.symmetric(horizontal: 15, vertical: DeviceScaler().scale(7)),
-                            decoration: BoxDecoration(
-                              color: Palette.primary,
-                              borderRadius: BorderRadius.circular(3),
-                            ),
-                            child: Text(
-                              'Send Privates',
-                              style: TextStyle(
-                                  color: Palette.white,
-                                  fontSize: DeviceScaler().scaleFont(10)
-                              ),
-                            ),
-                          ),
-                        ],
-                      ),
-                    ),
-                    Container(
-                      width: 150,
-                      padding: EdgeInsets.symmetric(vertical: DeviceScaler().scale(10)),
-                      margin: const EdgeInsets.symmetric(horizontal: 10),
-                      decoration: BoxDecoration(
-                          color: Palette.white,
-                          borderRadius: BorderRadius.circular(10),
-                          boxShadow: const [
-                            BoxShadow(
-                                blurRadius: 50,
-                                offset: Offset(0,0),
-                                color: Palette.lightShadowGrey
-                            )
-                          ]
-                      ),
-                      child: Column(
-                        mainAxisAlignment: MainAxisAlignment.end,
-                        children: [
-                          Image.asset("assets/png/big-photo2.png"),
-                          Text(
-                            "Angel Freeman",
-                            style: TextStyle(
-                              color: Palette.black,
-                              fontSize: DeviceScaler().scaleFont(14),
-                              fontWeight: FontWeight.w500,
-                            ),
-                          ),
-                          SizedBox(height: DeviceScaler().scale(3),),
-                          Text(
-                            "Actress , Life coach, \nphilanthropist",
-                            textAlign: TextAlign.center,
-                            style: TextStyle(
-                              color: Palette.grey,
-                              fontSize: DeviceScaler().scaleFont(10),
-                              fontWeight: FontWeight.w500,
-                            ),
-                          ),
-                          SizedBox(height: DeviceScaler().scale(5),),
-                          Container(
-                            padding: EdgeInsets.symmetric(horizontal: 15, vertical: DeviceScaler().scale(7)),
-                            decoration: BoxDecoration(
-                              color: Palette.primary,
-                              borderRadius: BorderRadius.circular(3),
-                            ),
-                            child: Text(
-                              'Send Privates',
-                              style: TextStyle(
-                                  color: Palette.white,
-                                  fontSize: DeviceScaler().scaleFont(10)
-                              ),
-                            ),
-                          ),
-                        ],
-                      ),
-                    ),
-                    Container(
-                      width: 150,
-                      padding: EdgeInsets.symmetric(vertical: DeviceScaler().scale(10)),
-                      margin: const EdgeInsets.symmetric(horizontal: 10),
-                      decoration: BoxDecoration(
-                          color: Palette.white,
-                          borderRadius: BorderRadius.circular(10),
-                          boxShadow: const [
-                            BoxShadow(
-                                blurRadius: 50,
-                                offset: Offset(0,0),
-                                color: Palette.lightShadowGrey
-                            )
-                          ]
-                      ),
-                      child: Column(
-                        mainAxisAlignment: MainAxisAlignment.end,
-                        children: [
-                          Image.asset("assets/png/big-photo.png"),
-                          Text(
-                            "Billy Morgan",
-                            style: TextStyle(
-                              color: Palette.black,
-                              fontSize: DeviceScaler().scaleFont(14),
-                              fontWeight: FontWeight.w500,
-                            ),
-                          ),
-                          SizedBox(height: DeviceScaler().scale(3),),
-                          Text(
-                            "Actress , Life coach, \nphilanthropist",
-                            textAlign: TextAlign.center,
-                            style: TextStyle(
-                              color: Palette.grey,
-                              fontSize: DeviceScaler().scaleFont(10),
-                              fontWeight: FontWeight.w500,
-                            ),
-                          ),
-                          SizedBox(height: DeviceScaler().scale(5),),
-                          Container(
-                            padding: EdgeInsets.symmetric(horizontal: 15, vertical: DeviceScaler().scale(7)),
-                            decoration: BoxDecoration(
-                              color: Palette.primary,
-                              borderRadius: BorderRadius.circular(3),
-                            ),
-                            child: Text(
-                              'Send Privates',
-                              style: TextStyle(
-                                  color: Palette.white,
-                                  fontSize: DeviceScaler().scaleFont(10)
-                              ),
-                            ),
-                          ),
-                        ],
-                      ),
-                    ),
+                  children: const [
+                    AddUserWidget(imageUrl: "assets/png/big-photo.png",),
+                    AddUserWidget(imageUrl: "assets/png/big-photo2.png",),
+                    AddUserWidget(imageUrl: "assets/png/big-photo.png",),
+                    AddUserWidget(imageUrl: "assets/png/big-photo2.png",),
+
                   ],
                 ),
               ),
@@ -260,5 +101,7 @@ class HomeView extends StatelessWidget {
     );
   }
 }
+
+
